Support Linux

+++ Src/ChromeDriverUpdater.Tests/UpdateTest.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 using NUnit.Framework;
 
@@ -7,10 +8,39 @@ namespace ChromeDriverUpdater.Tests
     public class UpdateTest
     {
         [Test]
+        public void LinuxTest()
+        {
+            if(File.Exists("chromedriver"))
+            {
+                File.Delete("chromedriver");
+            }
+
+            File.Copy("chromedriver_linux64", "chromedriver");
+            
+            Updater updater = new Updater();
+            try
+            {
+                updater.Update("chromedriver");
+            } 
+            catch(UpdateFailException exc)
+            {
+                Console.WriteLine(exc);
+                // In Github action, Chrome is not downloaded.
+                // Or cannot access to registry.
+                Assert.IsTrue(exc.ErrorCode == Models.ErrorCode.ChromeNotInstalled);
+            }
+            Assert.IsTrue(true);
+        }
+
+        [Test]
         public void ChromeDriverUpdateTest()
         {
-            File.Delete("chromedriver.exe");
-            File.Copy("chromedriver_win32.exe", "chromedriver.exe");
+            if(File.Exists("chromedriver.exe"))
+            {
+                File.Delete("chromedriver.exe");
+            }
+            
+            File.Copy("chromedriver_old.exe", "chromedriver.exe");
 
             Updater updater = new Updater();
             try
@@ -33,7 +63,7 @@ namespace ChromeDriverUpdater.Tests
             try
             {
                 Updater updater = new Updater();
-                updater.Update("wrong.exe");
+                updater.Update("wrong");
             }
             catch(UpdateFailException exc)
             {
